feat: reorganize styles and improve UserTable and LoginPage components for better maintainability
