+++ src/components/product/ProductTable.jsx
@@ -3,6 +3,7 @@ import {useEffect} from "react";
 import {useProductTable} from "../../hooks/product/useProductTable.jsx";
 import {Box, Paper, TextField} from "@mui/material";
 import {DataGrid} from "@mui/x-data-grid";
+import {userTableStyles} from "../../styles/js/UserTable.js";
 
 export const ProductTable = () => {
     const {products, handleGetProducts} = useProduct();
@@ -13,8 +14,8 @@ export const ProductTable = () => {
     }, []);
 
     return (
-        <Paper sx={{p: 2, height: "100%"}}>
-            <Box sx={{mb: 2, display: 'flex', justifyContent: 'space-between', alignItems: 'flex-start', gap: 2}}>
+        <Paper sx={userTableStyles.paper}>
+            <Box sx={userTableStyles.searchContainer}>
                 <TextField
                     fullWidth
                     label="Buscar productos..."
@@ -23,7 +24,7 @@ export const ProductTable = () => {
                     onChange={(e) => setSearchText(e.target.value)}
                 />
             </Box>
-            <Box sx={{height: 600, width: "100%"}}>
+            <Box sx={userTableStyles.tableContainer}>
                 <DataGrid
                     rows={filteredProducts}
                     columns={columns}
@@ -35,22 +36,7 @@ export const ProductTable = () => {
                     pageSizeOptions={[5, 10, 25, 50]}
                     disableRowSelectionOnClick
                     loading={!products}
-                    sx={{
-                        "& .MuiDataGrid-root": {
-                            border: "none",
-                        },
-                        "& .MuiDataGrid-cell": {
-                            borderBottom: "none",
-                        },
-                        "& .MuiDataGrid-columnHeaders": {
-                            backgroundColor: "#f5f5f5",
-                            color: "#000",
-                            fontSize: 16,
-                        },
-                        "& .MuiDataGrid-virtualScroller": {
-                            backgroundColor: "#fff",
-                        },
-                    }}
+                    sx={userTableStyles.dataGrid}
                 />
             </Box>
         </Paper>
