+++ src/styles/css/LoginPage.css
@@ -1,5 +1,5 @@
 /* Import Font Awesome */
-@import '@fortawesome/fontawesome-free/css/all.min.css';
+@import '../../../node_modules/@fortawesome/fontawesome-free/css/all.min.css';
 
 /* Login Page Styles for Fruit Store POS */
